fix: P4 真实 PG 验证通过——定位并修复 PG 方言类型映射 gap：INTEGER=INT4(32位)，version 列读取/写入按 i32(align SQLite INTEGER 8字节→PG 4字节)。pg_ds_entry_roundtrip 在真实 PostgreSQL(5433) 通过：连接+迁移14表+create_dataset/get_dataset/add_entry(snapshot去重事务)/list_entries+content_hash 往返全绿

+++ src/store/pg.rs
@@ -222,7 +222,8 @@ impl PgStore {
         )
         .bind(&entry.dataset_id)
         .bind(&entry.entry_id)
-        .bind(entry.version as i64)
+        // PG `INTEGER` = INT4，按 i32 bind（与入读一致）
+        .bind(entry.version as i32)
         .bind(
             entry
                 .status
@@ -285,7 +286,8 @@ fn map_entry_row(
     let entry = crate::model::entry::RuleEntry {
         dataset_id: r.get("dataset_id"),
         entry_id: r.get("entry_id"),
-        version: r.get::<i64, _>("version") as u32,
+        // PG `INTEGER` = INT4（32 位），需按 i32 读取再转 u32（对齐 entries.version 类型）
+        version: r.get::<i32, _>("version") as u32,
         status: status.map(|s| decode_opt(Some(s))).transpose()?.flatten(),
         provenance: decode_opt(r.get("provenance"))?.unwrap(),
         domain: r.get("domain"),
